Clean rounded floats.

+++ apps/profile/models.py
@@ -1208,8 +1208,8 @@ class PaymentHistory(models.Model):
             end_time = start_date + datetime.timedelta(days=31)
             end_date = datetime.datetime(end_time.year, end_time.month, 1) - datetime.timedelta(seconds=1)
             if end_date > now:
-                payments = {'avg': 0, 'sum': round(this_mtd_sum / (max(1, last_mtd_sum) / float(max(1, last_month_sum)))), 
-                            'count': round(this_mtd_count / (max(1, last_mtd_count) / float(max(1, last_month_count))))}
+                payments = {'avg': 0, 'sum': int(round(this_mtd_sum / (max(1, last_mtd_sum) / float(max(1, last_month_sum))))), 
+                            'count': int(round(this_mtd_count / (max(1, last_mtd_count) / float(max(1, last_month_count)))))}
                 _counter(start_date, end_date, payments=payments)
             else:
                 count = _counter(start_date, end_date)
